+++ apptive/src/main/java/com/backend/apptive/service/impl/UserServiceImpl.java
@@ -1,0 +1,4 @@
+package com.backend.apptive.service.impl;
+
+public class UserServiceImpl {
+}
